Update archive.py
Better fix for gallery detection. Down to <10 files in several thousand not being picked up, of which most are due to unsupported zip compression in the zipfile module. The remainder can be fixed with a change to the check_archive function but I can't find it in this codebase and it's probably not really worth the fuss I put into it anyway for so few archives...